+++ example/main.dart
@@ -4,32 +4,32 @@ void basis() {
   print("\n=========== Basic Usage ===========");
 
   final count = signal(1);
-  final doubleCount = computed((_) => count() * 2);
+  final doubleCount = computed((_) => count.value * 2);
 
   effect(() {
-    print("Count is: ${count()}");
+    print("Count is: ${count.value}");
   }); // Count is: 1
 
-  print(doubleCount()); // 2
+  print(doubleCount.value); // 2
 
-  count(2); // Count is: 2
+  count.value = 2; // Count is: 2
 
-  print(doubleCount()); // 4
+  print(doubleCount.value); // 4
 }
 
 void scope() {
   print("\n=========== Effect Scope ===========");
 
   final count = signal(1);
-  final stop = effectScope(() {
+  final scope = effectScope(() {
     effect(() {
-      print("Count is: ${count()}");
+      print("Count is: ${count.value}");
     }); // Count is: 1
   });
 
-  count(2); // Count is: 2
-  stop();
-  count(3); // Not printed
+  count.value = 2; // Count is: 2
+  scope.dispose();
+  count.value = 3; // Not printed
 }
 
 void main() {