+++ lib/preset_developer.dart
@@ -1,0 +1,12 @@
+/// Alien Signal preset for developers.
+///
+/// You can use the presets to further customize.
+library;
+
+export 'src/preset.dart'
+    show
+        PresetComputed,
+        PresetEffect,
+        PresetEffectScope,
+        PresetWritableSignal,
+        system;