+++ lib/system.dart
@@ -1,1 +1,4 @@
+/// Alien reactive system library.
+library;
+
 export 'src/system.dart' hide Stack;