+++ test/computed_test.dart
@@ -4,46 +4,46 @@ import 'package:test/test.dart';
 void main() {
   test("should correctly propagate changes through computed signals", () {
     final s = signal(0);
-    final c1 = computed((_) => s() % 2);
-    final c2 = computed((_) => c1());
-    final c3 = computed((_) => c2());
+    final c1 = computed((_) => s.value % 2);
+    final c2 = computed((_) => c1.value);
+    final c3 = computed((_) => c2.value);
 
-    c3();
-    s(1);
-    c2();
-    s(3);
+    c3.value;
+    s.value = 1;
+    c2.value;
+    s.value = 3;
 
-    expect(c3(), 1);
+    expect(c3.value, 1);
   });
 
   test("should propagate updated source value through chained computations",
       () {
     final s = signal(0);
-    final a = computed((_) => s());
-    final b = computed((_) => a() % 2);
-    final c = computed((_) => s());
-    final d = computed((_) => b() + c());
+    final a = computed((_) => s.value);
+    final b = computed((_) => a.value % 2);
+    final c = computed((_) => s.value);
+    final d = computed((_) => b.value + c.value);
 
-    expect(d(), 0);
-    s(2);
-    expect(d(), 2);
+    expect(d.value, 0);
+    s.value = 2;
+    expect(d.value, 2);
   });
 
   test("should handle flags are indirectly updated during checkDirty", () {
     final a = signal(false);
-    final b = computed((_) => a());
+    final b = computed((_) => a.value);
     final c = computed((_) {
-      b();
+      b.value;
       return 0;
     });
     final d = computed((_) {
-      c();
-      return b();
+      c.value;
+      return b.value;
     });
 
-    expect(d(), false);
-    a(true);
-    expect(d(), true);
+    expect(d.value, false);
+    a.value = true;
+    expect(d.value, true);
   });
 
   test("should not update if the signal value is reverted", () {
@@ -51,14 +51,14 @@ void main() {
     final s = signal(0);
     final c = computed((_) {
       times++;
-      return s();
+      return s.value;
     });
 
-    c();
+    c.value;
     expect(times, 1);
-    s(1);
-    s(0);
-    c();
+    s.value = 1;
+    s.value = 0;
+    c.value;
     expect(times, 1);
   });
 }